Added settings activity. Added modifiable behaviour of log button. Changed DBHelper for ContentProvider. Some lists now use cursor loader. Graphs are no longer displayed when marking more. Added floating button to main screen.

+++ app/src/main/java/com/example/marekulip/droidsor/sensorlogmanager/SensorData.java
@@ -51,7 +51,7 @@ public class SensorData {
         return cv;
     }
 
-    public ContentValues getInsertableFormat(int logId){
+    public ContentValues getInsertableFormat(long logId){
         ContentValues cv = getInsertableFormat();
         cv.put(SensorDataTable.LOG_ID,logId);
         return cv;
